requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
SQLAlchemy==2.0.43
Jinja2==3.1.6
pydantic==2.11.7
python-dotenv>=1.0.1
openai-whisper
torch
loguru>=0.7.0
deepgram-sdk>=6.0.1
pipecat-ai==0.0.108
